+++ railties/lib/rails/application/configuration.rb
@@ -370,6 +370,10 @@ module Rails
           end
         when "8.2"
           load_defaults "8.1"
+
+          if respond_to?(:active_model)
+            active_model.secure_password_algorithm = :argon2
+          end
         else
           raise "Unknown version #{target_version.to_s.inspect}"
         end